# specification-module/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Store API - OpenAPI 3.0
  description: |-
    This is a Store API Server
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://store-api.swagger.io/api/v3
...
paths:
...
  /sku:
    put:
      tags:
        - sku
      summary: Update an existing SKU
      description: Update an existing SKU by Id
      operationId: updateSku
      requestBody:
        description: Update an existent SKU in the app
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkuDto'
          application/xml:
            schema:
              $ref: '#/components/schemas/SkuDto'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SkuDto'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkuResponseDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/SkuResponseDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: SKU not found
        '405':
          description: Validation exception
      security:
        - bearerAuth: []
    post:
      tags:
        - sku
      summary: Add a new SKU to the app
      description: Add a new SKU to the app
      operationId: addSku
      requestBody:
        description: Create a new SKU to the app
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkuDto'
          application/xml:
            schema:
              $ref: '#/components/schemas/SkuDto'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SkuDto'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkuResponseDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/SkuResponseDto'
        '405':
          description: Invalid input
      security:
        - bearerAuth: []
  /sku/findByStatus:
    get:
      tags:
        - sku
      summary: Finds SKU by status
      description: Multiple status values can be provided with comma separated strings
      operationId: findSkuByStatus
      parameters:
        - name: status
          in: query
          description: Status values that need to be considered for filter
          required: false
          explode: true
          schema:
            type: string
            default: ACTIVE
            enum:
              - ACTIVE
              - DEACTIVE
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SkuResponseDto'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SkuResponseDto'
        '400':
          description: Invalid status value
      security:
        - bearerAuth: []
  /sku/{skuId}:
    get:
      tags:
        - sku
      summary: Find SKU by ID
      description: Returns a single SKU
      operationId: getSkuById
      parameters:
        - name: skuId
          in: path
          description: ID of SKU to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkuResponseDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/SkuResponseDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: SKU not found
      security:
        - bearerAuth: []
    post:
      tags:
        - sku
      summary: Updates a SKU in the app with form data
      description: ''
      operationId: updateSkuWithForm
      parameters:
        - name: skuId
          in: path
          description: ID of SKU to return
          required: true
          schema:
            type: integer
            format: int64
        - name: skuName
          in: query
          description: Name of Product that needs to be updated
          schema:
            type: string
        - name: status
          in: query
          description: Status of SKU that needs to be updated
          schema:
            type: string
      responses:
        '405':
          description: Invalid input
      security:
        - bearerAuth: []
    delete:
      tags:
        - sku
      summary: Deletes a SKU
      description: delete a SKU
      operationId: deleteSku
      parameters:
        - name: api_key
          in: header
          description: ''
          required: false
          schema:
            type: string
        - name: skuId
          in: path
          description: SKU id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid SKU value
      security:
        - bearerAuth: []
  /sku/{skuId}/uploadImage:
    post:
      tags:
        - sku
      summary: uploads an image
      description: ''
      operationId: uploadSkuFile
      parameters:
        - name: skuId
          in: path
          description: ID of sku to update
          required: true
          schema:
            type: integer
            format: int64
        - name: additionalMetadata
          in: query
          description: Additional Metadata
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    ProductDto:
      required:
        - productName
      type: object
      properties:
        productId:
          type: integer
          format: int64
          example: 10
        productName:
          type: string
          example: Dot 3, Dot 4
        photoUrl:
          type: string
        brandId:
          type: integer
          format: int64
          example: 10
        categoryId:
          type: integer
          format: int64
          example: 10
        status:
          type: string
          description: Product status in the app
          enum:
            - ACTIVE
            - DEACTIVE
      xml:
        name: ProductDto
...
    SkuDto:
      required:
        - skuName
      type: object
      properties:
        skuId:
          type: integer
          format: int64
          example: 10
        skuName:
          type: string
          example: 2L, 4L
        photoUrl:
          type: string
        productId:
          type: integer
          format: int64
          example: 10
        status:
          type: string
          description: SKU status in the app
          enum:
            - ACTIVE
            - DEACTIVE
      xml:
        name: SkuDto
    SkuResponseDto:
      required:
        - skuName
        - quantity
      type: object
      properties:
        skuId:
          type: integer
          format: int64
          example: 10
        skuName:
          type: string
          example: 2L, 4L
        photoUrl:
          type: string
        product:
          $ref: '#/components/schemas/ProductDto'
        status:
          type: string
          description: SKU status in the app
          enum:
            - ACTIVE
            - DEACTIVE
      xml:
        name: skuResponse
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: '##default'
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT